JC- Added in new assets and working on arrows firing & movment, added CD/CR to islands and arrows

+++ GD4SFMLGameWorld/Player.hpp
@@ -18,7 +18,7 @@ class CommandQueue;
 class Player
 {
 public:
-	Player(sf::TcpSocket* socket, sf::Int32 identifier, const KeyBinding* binding);
+	Player(sf::TcpSocket* socket, sf::Uint8 identifier, const KeyBinding* binding);
 
 	void handleEvent(const sf::Event& event, CommandQueue& commands);
 	void handleRealtimeInput(CommandQueue& commands);
